+++ tests/pytests/test_mex_drivers.py
@@ -10,519 +10,6 @@ import ale
 
 from ale.drivers.mex_drivers import MexHrscPds3NaifSpiceDriver, MexHrscIsisLabelNaifSpiceDriver, MexSrcPds3NaifSpiceDriver 
 
-@pytest.fixture()
-def usgscsm_compare_dict():
-    return {
-    "h5270_0000_ir2" : {
-        "usgscsm" : {
-          "radii": {
-            "semimajor": 3396.19,
-            "semiminor": 3376.2,
-            "unit": "km"
-              },
-              "sensor_position": {
-                "positions": [
-                  [
-                    711902.968354,
-                    3209827.60790571,
-                    1748326.86116295
-                  ],
-                  [
-                    727778.89367768,
-                    3287885.02005966,
-                    1594882.70156054
-                  ],
-                  [
-                    743098.34408384,
-                    3361360.97987664,
-                    1439236.15929773
-                  ],
-                  [
-                    757817.60768561,
-                    3430175.96634995,
-                    1281609.7397002
-                  ],
-                  [
-                    771895.91691839,
-                    3494268.82029183,
-                    1122231.09675971
-                  ],
-                  [
-                    785295.6426853,
-                    3553596.86562762,
-                    961331.02292412
-                  ]
-                ],
-                "velocities": [
-                  [
-                    396.19391017,
-                    1971.70523609,
-                    -3738.08862116
-                  ],
-                  [
-                    383.09225613,
-                    1860.35892297,
-                    -3794.8312507
-                  ],
-                  [
-                    368.88320115,
-                    1746.8383078,
-                    -3846.19171074
-                  ],
-                  [
-                    353.63635876,
-                    1631.58973504,
-                    -3892.0182367
-                  ],
-                  [
-                    337.42645506,
-                    1515.06674438,
-                    -3932.20958309
-                  ],
-                  [
-                    320.33289561,
-                    1397.72243165,
-                    -3966.71239887
-                  ]
-                ],
-                "unit": "m"
-              },
-              "sun_position": {
-                "positions": [
-                  [
-                    2.05222074E+11,
-                    1.19628335E+11,
-                    5.02349719E+10
-                  ]
-                ],
-                "velocities": [
-                  [
-                    8468758.54,
-                    -14528713.8,
-                    8703.55212
-                  ]
-                ],
-                "unit": "m"
-              },
-              "sensor_orientation": {
-                "quaternions": [
-                  [
-                    -0.09146728,
-                    -0.85085751,
-                    0.51357522,
-                    0.06257586
-                  ],
-                  [
-                    -0.09123532,
-                    -0.83858882,
-                    0.53326329,
-                    0.0638371
-                  ],
-                  [
-                    -0.09097193,
-                    -0.82586685,
-                    0.55265188,
-                    0.06514567
-                  ],
-                  [
-                    -0.09050679,
-                    -0.81278131,
-                    0.57165363,
-                    0.06638667
-                  ],
-                  [
-                    -0.08988786,
-                    -0.79935128,
-                    0.59024631,
-                    0.06757961
-                  ],
-                  [
-                    -0.08924306,
-                    -0.78551905,
-                    0.60849234,
-                    0.06879366
-                  ]
-                ]
-              },
-              "detector_sample_summing": 1,
-              "detector_line_summing": 1,
-              "focal_length_model": {
-                "focal_length": 174.82
-              },
-              "detector_center": {
-                "line": 0.0,
-                "sample": 2592.0
-              },
-              "starting_detector_line": 0,
-              "starting_detector_sample": 0,
-              "focal2pixel_lines": [
-                -7113.11359717265,
-                0.062856784318668,
-                142.857129028729
-              ],
-              "focal2pixel_samples": [
-                -0.778052433438109,
-                -142.857129028729,
-                0.062856784318668
-              ],
-              "optical_distortion": {
-                "radial": {
-                  "coefficients": [
-                    0.0,
-                    0.0,
-                    0.0
-                  ]
-                }
-              },
-              "image_lines": 400,
-              "image_samples": 1288,
-              "name_platform": "MARS EXPRESS",
-              "name_sensor": "HIGH RESOLUTION STEREO CAMERA",
-              "reference_height": {
-                "maxheight": 1000,
-                "minheight": -1000,
-                "unit": "m"
-              },
-              "name_model": "USGS_ASTRO_LINE_SCANNER_SENSOR_MODEL",
-              "interpolation_method": "lagrange",
-              "line_scan_rate": [
-                [
-                  0.5,
-                  -94.88182842731476,
-                  0.012800790786743165
-                ],
-                [
-                  15086.5,
-                  101.82391116023064,
-                  0.013227428436279297
-                ]
-              ],
-              "starting_ephemeris_time": 255744592.07217148,
-              "center_ephemeris_time": 255744693.90931007,
-              "t0_ephemeris": -101.83713859319687,
-              "dt_ephemeris": 40.734855437278746,
-              "t0_quaternion": -101.83713859319687,
-              "dt_quaternion": 40.734855437278746
-              },
-
-        "isis" :
-        {
-  "CameraVersion": 1,
-  "NaifKeywords": {
-    "BODY499_RADII": [
-      3396.19,
-      3396.19,
-      3376.2
-    ],
-    "BODY_FRAME_CODE": 10014,
-    "BODY_CODE": 499,
-    "INS-41210_FOV_FRAME": "MEX_HRSC_HEAD",
-    "FRAME_-41210_NAME": "MEX_HRSC_HEAD",
-    "INS-41210_CK_TIME_TOLERANCE": 1.0,
-    "TKFRAME_-41210_AXES": [
-      1.0,
-      2.0,
-      3.0
-    ],
-    "TKFRAME_-41210_SPEC": "ANGLES",
-    "FRAME_-41210_CLASS": 4.0,
-    "INS-41210_FOV_ANGULAR_SIZE": [
-      0.2,
-      0.659734
-    ],
-    "INS-41210_OD_K": [
-      0.0,
-      0.0,
-      0.0
-    ],
-    "INS-41210_F/RATIO": 5.6,
-    "INS-41210_PLATFORM_ID": -41000.0,
-    "TKFRAME_-41210_ANGLES": [
-      -0.334,
-      0.0101,
-      0.0
-    ],
-    "INS-41210_SPK_TIME_BIAS": 0.0,
-    "FRAME_-41210_CENTER": -41.0,
-    "TKFRAME_-41210_UNITS": "DEGREES",
-    "INS-41210_BORESIGHT": [
-      0.0,
-      0.0,
-      175.0
-    ],
-    "INS-41210_CK_TIME_BIAS": 0.0,
-    "FRAME_-41210_CLASS_ID": -41210.0,
-    "INS-41210_IFOV": 4e-05,
-    "INS-41210_FOV_BOUNDARY_CORNERS": [
-      18.187,
-      60.0641,
-      175.0,
-      18.1281,
-      -60.0399,
-      175.0,
-      -18.1862,
-      -60.0435,
-      175.0,
-      -18.142
-    ],
-    "INS-41210_FOV_SHAPE": "RECTANGLE",
-    "TKFRAME_-41210_RELATIVE": "MEX_HRSC_BASE",
-    "INS-41210_PIXEL_PITCH": 0.007,
-    "INS-41210_FOCAL_LENGTH": 175.0,
-    "BODY499_POLE_DEC": [
-      52.8865,
-      -0.0609,
-      0.0
-    ],
-    "BODY499_POLE_RA": [
-      317.68143,
-      -0.1061,
-      0.0
-    ],
-    "BODY499_PM": [
-      176.63,
-      350.89198226,
-      0.0
-    ],
-    "INS-41218_ITRANSL": [
-      -7113.11359717265,
-      0.062856784318668,
-      142.857129028729
-    ],
-    "INS-41218_ITRANSS": [
-      -0.778052433438109,
-      -142.857129028729,
-      0.062856784318668
-    ],
-    "INS-41218_FOV_SHAPE": "RECTANGLE",
-    "INS-41218_PIXEL_SIZE": [
-      7.0,
-      7.0
-    ],
-    "INS-41218_CK_REFERENCE_ID": 1.0,
-    "INS-41218_FOV_FRAME": "MEX_HRSC_HEAD",
-    "INS-41218_CCD_CENTER": [
-      2592.5,
-      0.5
-    ],
-    "INS-41218_CK_FRAME_ID": -41001.0,
-    "INS-41218_F/RATIO": 5.6,
-    "INS-41218_PIXEL_SAMPLES": 5184.0,
-    "INS-41218_BORESIGHT_SAMPLE": 2592.5,
-    "INS-41218_FILTER_BANDWIDTH": 90.0,
-    "INS-41218_BORESIGHT_LINE": 0.0,
-    "INS-41218_PIXEL_LINES": 1.0,
-    "INS-41218_FOCAL_LENGTH": 174.82,
-    "INS-41218_FOV_ANGULAR_SIZE": [
-      0.2,
-      4e-05
-    ],
-    "INS-41218_FILTER_BANDCENTER": 970.0,
-    "INS-41218_TRANSX": [
-      0.016461898406507,
-      -0.006999999322408,
-      3.079982431615e-06
-    ],
-    "INS-41218_TRANSY": [
-      49.7917927568053,
-      3.079982431615e-06,
-      0.006999999322408
-    ],
-    "INS-41218_FOV_BOUNDARY_CORNERS": [
-      18.1982,
-      49.9121,
-      175.0,
-      18.1982,
-      49.9051,
-      175.0,
-      -18.1693,
-      49.8901,
-      175.0,
-      -18.1693
-    ],
-    "INS-41218_BORESIGHT": [
-      0.0151,
-      49.9039,
-      175.0
-    ],
-    "INS-41218_IFOV": 4e-05
-  },
-  "InstrumentPointing": {
-    "TimeDependentFrames": [
-      -41001,
-      1
-    ],
-    "CkTableStartTime": 255744599.02748,
-    "CkTableEndTime": 255744635.91477,
-    "CkTableOriginalSize": 3,
-    "EphemerisTimes": [
-      255744599.02748,
-      255744623.61901,
-      255744635.91477
-    ],
-    "Quaternions": [
-      [
-        -0.34147103206303764,
-        0.46006200041554185,
-        -0.48264106492774883,
-        -0.6624183666542334
-      ],
-      [
-        -0.34862899148129517,
-        0.4555408857335137,
-        -0.47327265910130095,
-        -0.668545673735942
-      ],
-      [
-        -0.3521802679309037,
-        0.45323805476596757,
-        -0.46855266563769715,
-        -0.6715673637959837
-      ]
-    ],
-    "AngularVelocity": [
-      [
-        0.00035176331113592204,
-        0.0010154650024473103,
-        0.00038771759244781866
-      ],
-      [
-        0.00035242855802833725,
-        0.0010149701470475953,
-        0.0003878218830533074
-      ],
-      [
-        0.0003502620823697415,
-        0.001017194110775444,
-        0.00038476436104443903
-      ]
-    ],
-    "ConstantFrames": [
-      -41210,
-      -41200,
-      -41000,
-      -41001
-    ],
-    "ConstantRotation": [
-      -0.9999999844629888,
-      1.027590578527487e-06,
-      0.00017627525841189352,
-      1.2246232944813223e-16,
-      -0.9999830090976747,
-      0.00582936668603668,
-      0.0001762782535384808,
-      0.0058293665954657434,
-      0.9999829935609271
-    ]
-  },
-  "BodyRotation": {
-    "TimeDependentFrames": [
-      10014,
-      1
-    ],
-    "CkTableStartTime": 255744599.02748,
-    "CkTableEndTime": 255744635.91477,
-    "CkTableOriginalSize": 2,
-    "EphemerisTimes": [
-      255744599.02748,
-      255744635.91477
-    ],
-    "Quaternions": [
-      [
-        -0.6525755651363003,
-        -0.0231514239139282,
-        0.3174415084289179,
-        -0.6876336467074378
-      ],
-      [
-        -0.6534739684048748,
-        -0.022736404778153148,
-        0.31747150360998055,
-        -0.68677993048033
-      ]
-    ],
-    "AngularVelocity": [
-      [
-        3.1623981615137114e-05,
-        -2.8803031775991542e-05,
-        5.6520727317788564e-05
-      ],
-      [
-        3.1623981615032794e-05,
-        -2.8803031777148914e-05,
-        5.6520727317257115e-05
-      ]
-    ]
-  },
-  "InstrumentPosition": {
-    "SpkTableStartTime": 255744599.02748,
-    "SpkTableEndTime": 255744635.91477,
-    "SpkTableOriginalSize": 3,
-    "EphemerisTimes": [
-      255744599.02748,
-      255744623.61901,
-      255744635.91477
-    ],
-    "Positions": [
-      [
-        3508.767882205483,
-        -1180.0905787748716,
-        -404.65806593586274
-      ],
-      [
-        3509.658413803481,
-        -1143.4324359444558,
-        -502.60294631877633
-      ],
-      [
-        3509.443153282348,
-        -1124.8866548757137,
-        -551.4851113671589
-      ]
-    ],
-    "Velocities": [
-      [
-        0.07204008324341267,
-        1.4787375673363454,
-        -3.987265079143158
-      ],
-      [
-        0.0003930097236827663,
-        1.5024971609204454,
-        -3.978142968522354
-      ],
-      [
-        -0.03540185319249475,
-        1.5140837760766446,
-        -3.9728346761198696
-      ]
-    ]
-  },
-  "SunPosition": {
-    "SpkTableStartTime": 255744697.39357847,
-    "SpkTableEndTime": 255744697.39357847,
-    "SpkTableOriginalSize": 1,
-    "EphemerisTimes": [
-      255744697.39357847
-    ],
-    "Positions": [
-      [
-        99139014.8191443,
-        -200427816.19676697,
-        -94608347.50413091
-      ]
-    ],
-    "Velocities": [
-      [
-        21.199131882686167,
-        7.334313602125272,
-        2.7913538260189816
-      ]
-    ]
-  }
-}
-}}
-
 
 @pytest.fixture()
 def test_mex_src_kernels(scope="module", autouse=True):
@@ -552,10 +39,8 @@ def test_mex_src_load(test_mex_src_kernels):
 
 # Eventually all label/formatter combinations should be tested. For now, isis3/usgscsm and
 # pds3/isis will fail.
-@pytest.mark.parametrize("label,formatter", [('isis3','isis'), ('pds3', 'usgscsm'),
-                                              pytest.param('isis3','usgscsm', marks=pytest.mark.xfail),
-                                              pytest.param('pds3','isis', marks=pytest.mark.xfail),])
-def test_mex_load(test_mex_hrsc_kernels, formatter, usgscsm_compare_dict, label):
+@pytest.mark.parametrize("label", [('isis3'), ('pds3')])
+def test_mex_load(test_mex_hrsc_kernels, label):
     label_file = get_image_label('h5270_0000_ir2', label)
 
     with patch('ale.drivers.mex_drivers.MexHrscPds3NaifSpiceDriver.binary_ephemeris_times', \
@@ -564,20 +49,24 @@ def test_mex_load(test_mex_hrsc_kernels, formatter, usgscsm_compare_dict, label)
                new_callable=PropertyMock) as binary_exposure_durations, \
         patch('ale.drivers.mex_drivers.MexHrscPds3NaifSpiceDriver.binary_lines', \
                new_callable=PropertyMock) as binary_lines, \
-        patch('ale.drivers.mex_drivers.MexHrscIsisLabelNaifSpiceDriver.ephemeris_time', \
+        patch('ale.base.type_sensor.LineScanner.ephemeris_time', \
                new_callable=PropertyMock) as ephemeris_time, \
         patch('ale.drivers.mex_drivers.read_table_data', return_value=12345) as read_table_data, \
         patch('ale.drivers.mex_drivers.parse_table', return_value={'EphemerisTime': [255744599.02748165, 255744684.33197814, 255744684.34504557], \
                                                                    'ExposureTime': [0.012800790786743165, 0.012907449722290038, 0.013227428436279297], \
                                                                    'LineStart': [1, 6665, 6666]}) as parse_table:
 
-        ephemeris_time.return_value = [255744599.02748, 255744623.61901, 255744635.91477]
+        ephemeris_time.return_value = [255744599.02748165, 255744684.33197814, 255744684.34504557]
         binary_ephemeris_times.return_value = [255744599.02748165, 255744599.04028246, 255744795.73322123]
         binary_exposure_durations.return_value = [0.012800790786743165, 0.012800790786743165, 0.013227428436279297]
-        binary_lines.return_value = [0.5, 1.5, 15086.5]
+        binary_lines.return_value = [0.5, 6664.5, 6665.5]
 
-        usgscsm_isd = ale.load(label_file, props={'kernels': test_mex_hrsc_kernels}, formatter=formatter)
-        assert compare_dicts(usgscsm_isd, usgscsm_compare_dict['h5270_0000_ir2'][formatter]) == []
+        usgscsm_isd = ale.load(label_file, props={'kernels': test_mex_hrsc_kernels})
+        if label == "isis3":
+          compare_isd = get_isd('mexhrsc_isis')
+        else:
+          compare_isd = get_isd('mexhrsc')
+        assert compare_dicts(usgscsm_isd, compare_isd) == []
 
 # ========= Test mex pds3label and naifspice driver =========
 class test_mex_pds3_naif(unittest.TestCase):
@@ -588,9 +77,6 @@ class test_mex_pds3_naif(unittest.TestCase):
 
     def test_short_mission_name(self):
         assert self.driver.short_mission_name=='mex'
-
-    def test_odtk(self):
-        assert self.driver.odtk == [0.0, 0.0, 0.0]
 
     def test_ikid(self):
         with patch('ale.drivers.mex_drivers.spice.bods2c', return_value=12345) as bods2c:
@@ -726,7 +212,7 @@ class test_mex_isis3_naif(unittest.TestCase):
                                                                    'ExposureTime': [0.012800790786743165, 0.012907449722290038, 0.013227428436279297], \
                                                                    'LineStart': [1, 6665, 6666]}) as parse_table:
 
-             assert self.driver.line_scan_rate == ([1, 6665, 6666], [255744599.02748165, 255744684.33197814, 255744684.34504557], [0.012800790786743165, 0.012907449722290038, 0.013227428436279297])
+             assert self.driver.line_scan_rate == ([0.5, 6664.5, 6665.5], [-98.36609682440758, -13.06160032749176, -13.048532903194427], [0.012800790786743165, 0.012907449722290038, 0.013227428436279297])
 
     def test_ephemeris_stop_time(self):
         with patch('ale.drivers.mex_drivers.read_table_data', return_value=12345) as read_table_data, \
@@ -756,9 +242,6 @@ class test_mex_src_pds3_naif(unittest.TestCase):
 
     def test_short_mission_name(self):
         assert self.driver.short_mission_name=='mex'
-
-    def test_odtk(self):
-        assert self.driver.odtk == [0.0, 0.0, 0.0]
 
     def test_ikid(self):
         with patch('ale.drivers.mex_drivers.spice.bods2c', return_value=12345) as bods2c: